refactor: Move SeekDto to DTOs folder
