+++ src/main/java/tak/Seek.java
@@ -10,6 +10,8 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import static tak.Game.DEFAULT_SIZE;
+
+import tak.DTOs.SeekDto;
 import tak.utils.ConcurrentHashSet;
 import java.util.concurrent.locks.*;
